Fix: ImageBuffer blank constructor
The constructor uses a default value for initializing the pixels. The problem was that the value was supposed to be a raw RGBA value to pass to the `_decodePixel` method and we were using the value returned by `minPixelValues` instead.

The fix was simply to use a default raw RGBA value: [0, 0, 0, 0]